requirements.txt
Flask-SQLAlchemy==3.1.1
mysqlclient==2.2.3
Werkzeug==2.2.2
Flask-OAuthlib==0.9.6
python-dotenv==1.0.1